feat: закончить работу по 14 проектной

+++ controllers/cards.js
@@ -1,48 +1,66 @@
-const Card = require('../models/card');
+/* eslint no-underscore-dangle: ["error", { "allow": ["_id"] }] */
 
-module.exports.getCards = (req, res) => {
+const Card = require('../models/card');
+const NotFoundError = require('../utils/Errors/NotFoundError');
+const AuthError = require('../utils/Errors/AuthError');
+const BadRequestError = require('../utils/Errors/BadRequestError');
+const ForbiddenError = require('../utils/Errors/ForbiddenError');
+
+module.exports.getCards = (req, res, next) => {
+  if (!req.headers.token) {
+    throw new AuthError('Нет доступа, необходима авторизация');
+  }
   Card.find({})
     .then((cards) => res.status(200).send(cards))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
+  if (!req.headers.token) {
+    throw new AuthError('Нет доступа, необходима авторизация');
+  }
   const { name, link } = req.body;
-  const owner = req.user._id;
 
-  Card.create({ name, link, owner })
+  Card.create({ name, link, owner: req.user })
     .then((card) => res.status(201).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(400).send({ message: `Произошла ошибка ${err.message}` });
-        return;
+        return next(new BadRequestError(`Данные карточки невалидны ${err.message}`));
       }
-      res.status(500).send({ message: 'На сервере произошла ошибка' });
+      return next(err);
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params._id)
+module.exports.deleteCard = (req, res, next) => {
+  if (!req.headers.token) {
+    throw new AuthError('Нет доступа, необходима авторизация');
+  }
+  Card.findById(req.params._id)
     .then((card) => {
       if (!card) {
-        res.status(404).send({ message: 'Карточка не найдена' });
-        return;
+        return next(new NotFoundError('Карточка не найдена'));
       }
-      res.status(200).send({ message: 'Карточка удалена' });
+      if (card.owner.toString() !== req.user) {
+        return next(new ForbiddenError('Нет прав для удаления карточки'));
+      }
+      return Card.findByIdAndDelete(req.params._id)
+        .then(res.status(200).send({ message: 'Карточка удалена' }));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: `Произошла ошибка ${err.message}` });
-        return;
+        return next(new BadRequestError('Некорректный ID карточки'));
       }
-      res.status(500).send({ message: 'На сервере произошла ошибка' });
+      return next(err);
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
+  if (!req.headers.token) {
+    throw new AuthError('Нет доступа, необходима авторизация');
+  }
   Card.findByIdAndUpdate(
     req.params._id,
-    { $addToSet: { likes: req.user._id } },
+    { $addToSet: { likes: req.user } },
     { new: true },
   )
     .then((card) => {
@@ -54,17 +72,19 @@ module.exports.likeCard = (req, res) => {
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: `Произошла ошибка ${err.message}` });
-        return;
+        return next(new BadRequestError('Некорректный ID карточки'));
       }
-      res.status(500).send({ message: 'На сервере произошла ошибка' });
+      return next(err);
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
+  if (!req.headers.token) {
+    throw new AuthError('Нет доступа, необходима авторизация');
+  }
   Card.findByIdAndUpdate(
     req.params._id,
-    { $pull: { likes: req.user._id } },
+    { $pull: { likes: req.user } },
     { new: true },
   )
     .then((card) => {
@@ -76,11 +96,8 @@ module.exports.dislikeCard = (req, res) => {
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: `Произошла ошибка ${err.message}` });
-        return;
+        return next(new BadRequestError('Некорректный ID карточки'));
       }
-      res.status(500).send({ message: 'На сервере произошла ошибка' });
+      return next(err);
     });
 };
-
-/* eslint no-underscore-dangle: ["error", { "allow": ["_id"] }] */
